+++ src/node/gov/handlers/proposals.h
@@ -9,12 +9,70 @@
 #include "js/checks.h"
 #include "js/extensions/ccf/network.h"
 #include "js/extensions/ccf/node.h"
-#include "node/gov/api_types.h"
 #include "node/gov/api_version.h"
 #include "node/gov/handlers/helpers.h"
 
 namespace ccf::gov::endpoints
 {
+  namespace api
+  {
+    struct Proposal
+    {
+      ccf::ProposalId proposal_id;
+      ccf::MemberId proposer_id;
+      ccf::ProposalState proposal_state = ccf::ProposalState::OPEN;
+      size_t ballot_count = 0;
+      std::vector<ccf::MemberId> ballot_submitters;
+      std::optional<ccf::jsgov::Votes> final_votes = std::nullopt;
+      std::optional<ccf::jsgov::VoteFailures> vote_failures = std::nullopt;
+      std::optional<ccf::jsgov::Failure> failure = std::nullopt;
+    };
+    DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(Proposal);
+    DECLARE_JSON_REQUIRED_FIELDS_WITH_RENAMES(
+      Proposal,
+      proposal_id,
+      "proposalId",
+      proposer_id,
+      "proposerId",
+      proposal_state,
+      "proposalState",
+      ballot_count,
+      "ballotCount",
+      ballot_submitters,
+      "ballotSubmitters");
+    DECLARE_JSON_OPTIONAL_FIELDS_WITH_RENAMES(
+      Proposal,
+      final_votes,
+      "finalVotes",
+      vote_failures,
+      "voteFailures",
+      failure,
+      "failure");
+
+    struct ProposalList
+    {
+      std::vector<Proposal> value;
+    };
+    DECLARE_JSON_TYPE(ProposalList);
+    DECLARE_JSON_REQUIRED_FIELDS(ProposalList, value);
+
+    struct ProposalAction
+    {
+      std::string name;
+      nlohmann::json args = nullptr;
+    };
+    DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(ProposalAction);
+    DECLARE_JSON_REQUIRED_FIELDS(ProposalAction, name);
+    DECLARE_JSON_OPTIONAL_FIELDS(ProposalAction, args);
+
+    struct ProposalActions
+    {
+      std::vector<ProposalAction> actions;
+    };
+    DECLARE_JSON_TYPE(ProposalActions);
+    DECLARE_JSON_REQUIRED_FIELDS(ProposalActions, actions);
+  }
+
   namespace detail
   {
     struct ProposalSubmissionResult
@@ -350,58 +408,25 @@ namespace ccf::gov::endpoints
       }
     }
 
-    inline nlohmann::json convert_proposal_to_api_format(
+    inline api::Proposal convert_proposal_to_api_format(
       const ProposalId& proposal_id, const ccf::jsgov::ProposalInfo& summary)
     {
-      auto response_body = nlohmann::json::object();
-
-      response_body["proposalId"] = proposal_id;
-      response_body["proposerId"] = summary.proposer_id;
-      response_body["proposalState"] = summary.state;
-      response_body["ballotCount"] = summary.ballots.size();
-
-      auto ballot_submitters = nlohmann::json::array();
       std::vector<ccf::MemberId> submitter_ids;
       for (const auto& [member_id, _] : summary.ballots)
       {
         submitter_ids.push_back(member_id);
       }
       std::sort(submitter_ids.begin(), submitter_ids.end());
-      for (const auto& member_id : submitter_ids)
-      {
-        ballot_submitters.push_back(member_id);
-      }
-      response_body["ballotSubmitters"] = ballot_submitters;
 
-      std::optional<ccf::jsgov::Votes> votes = summary.final_votes;
-
-      if (votes.has_value())
-      {
-        auto final_votes = nlohmann::json::object();
-        for (const auto& [voter_id, vote_result] : *votes)
-        {
-          final_votes[voter_id.value()] = vote_result;
-        }
-        response_body["finalVotes"] = final_votes;
-      }
-
-      if (summary.vote_failures.has_value())
-      {
-        auto vote_failures = nlohmann::json::object();
-        for (const auto& [failer_id, failure] : *summary.vote_failures)
-        {
-          vote_failures[failer_id.value()] = failure;
-        }
-        response_body["voteFailures"] = vote_failures;
-      }
-
-      if (summary.failure.has_value())
-      {
-        auto failure = nlohmann::json::object();
-        response_body["failure"] = *summary.failure;
-      }
-
-      return response_body;
+      return {
+        proposal_id,
+        summary.proposer_id,
+        summary.state,
+        summary.ballots.size(),
+        std::move(submitter_ids),
+        summary.final_votes,
+        summary.vote_failures,
+        summary.failure};
     }
   }
 
@@ -835,18 +860,15 @@ namespace ccf::gov::endpoints
             ctx.tx.template ro<ccf::jsgov::ProposalInfoMap>(
               jsgov::Tables::PROPOSALS_INFO);
 
-          auto proposal_list = nlohmann::json::array();
+          api::ProposalList response_body;
           proposal_info_handle->foreach(
-            [&proposal_list](
+            [&response_body](
               const auto& proposal_id, const auto& proposal_info) {
-              auto api_proposal = detail::convert_proposal_to_api_format(
-                proposal_id, proposal_info);
-              proposal_list.push_back(api_proposal);
+              response_body.value.push_back(
+                detail::convert_proposal_to_api_format(
+                  proposal_id, proposal_info));
               return true;
             });
-
-          auto response_body = nlohmann::json::object();
-          response_body["value"] = proposal_list;
 
           ctx.rpc_ctx->set_response_json(response_body, HTTP_STATUS_OK);
           return;
@@ -890,6 +912,8 @@ namespace ccf::gov::endpoints
             return;
           }
 
+          (void)ccf::parse_json_safe(proposal.value())
+            .template get<api::ProposalActions>();
           ctx.rpc_ctx->set_response_status(HTTP_STATUS_OK);
           ctx.rpc_ctx->set_response_body(proposal.value());
           ctx.rpc_ctx->set_response_header(