+++ src/node/gov/handlers/transactions.h
@@ -6,11 +6,22 @@
 #include "ccf/json_handler.h"
 #include "ccf/tx_id.h"
 #include "ccf/tx_status.h"
-#include "node/gov/api_types.h"
 #include "node/gov/api_version.h"
 
 namespace ccf::gov::endpoints
 {
+  namespace api
+  {
+    struct Transaction
+    {
+      ccf::TxStatus status = ccf::TxStatus::Unknown;
+      ccf::TxID transaction_id;
+    };
+    DECLARE_JSON_TYPE(Transaction);
+    DECLARE_JSON_REQUIRED_FIELDS_WITH_RENAMES(
+      Transaction, status, "status", transaction_id, "transactionId");
+  }
+
   inline void init_transactions_handlers(ccf::BaseEndpointRegistry& registry)
   {
     auto get_transaction_status = [&](auto& ctx, ApiVersion api_version) {
@@ -69,10 +80,7 @@ namespace ccf::gov::endpoints
           }
 
           // Build response
-          auto body = nlohmann::json::object();
-
-          body["status"] = status;
-          body["transactionId"] = tx_id->to_str();
+          const api::Transaction body{status, tx_id.value()};
 
           ctx.rpc_ctx->set_response_json(body, HTTP_STATUS_OK);
           return;
@@ -128,10 +136,7 @@ namespace ccf::gov::endpoints
           }
 
           // Build response
-          auto body = nlohmann::json::object();
-
-          body["status"] = status;
-          body["transactionId"] = ccf::TxID{view, seqno};
+          const api::Transaction body{status, ccf::TxID{view, seqno}};
 
           ctx.rpc_ctx->set_response_json(body, HTTP_STATUS_OK);
           return;